Fix: Bean 오버라이드 문제
bean으로 등록되는 메서드명이 같아서 생기는 문제 해결
