+++ src/main/java/nbc/chillguys/nebulazone/infra/security/config/SecurityConfig.java
@@ -4,9 +4,8 @@ import java.util.List;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Profile;
-import org.springframework.core.annotation.Order;
 import org.springframework.http.HttpMethod;
+import org.springframework.security.config.Customizer;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
@@ -27,7 +26,6 @@ import nbc.chillguys.nebulazone.infra.security.filter.LoggingFilter;
 
 @Configuration
 @EnableWebSecurity
-@Profile("prod")
 @RequiredArgsConstructor
 public class SecurityConfig {
 	private final CustomAuthenticationEntryPoint entryPoint;
@@ -39,24 +37,9 @@ public class SecurityConfig {
 	private final HttpCookieOAuth2AuthorizationRequestRepository httpCookieOAuth2AuthorizationRequestRepository;
 
 	@Bean
-	@Order(1)
-	public SecurityFilterChain managementSecurityFilterChain(HttpSecurity http) throws Exception {
-		return http
-			.securityMatcher("/actuator/**")
-			.authorizeHttpRequests(auth -> auth
-				.requestMatchers("/actuator/health", "/actuator/prometheus").permitAll()
-				.anyRequest().denyAll()
-			)
-			.csrf(AbstractHttpConfigurer::disable)
-			.sessionManagement(session -> session.sessionCreationPolicy(SessionCreationPolicy.STATELESS))
-			.build();
-	}
-
-	@Bean
-	@Order(2)
-	public SecurityFilterChain appSecurityFilterChain(HttpSecurity http) throws Exception {
-		return http
-			.cors(cors -> cors.configurationSource(corsConfigurationSource()))
+	public SecurityFilterChain filterChain(HttpSecurity httpSecurity) throws Exception {
+		return httpSecurity
+			.cors(Customizer.withDefaults())
 			.sessionManagement(session -> session.sessionCreationPolicy(SessionCreationPolicy.STATELESS))
 			.csrf(AbstractHttpConfigurer::disable)
 			.authorizeHttpRequests(auth -> auth
@@ -68,6 +51,8 @@ public class SecurityConfig {
 					"/swagger-resources/**",
 					"/favicon.ico",
 					"/error",
+					"/actuator/**",
+					"/metrics/**",
 					"/ws/**",
 					"/ws",
 					"/chat/**",
@@ -87,10 +72,10 @@ public class SecurityConfig {
 					"/catalogs/**",
 					"/products/**"
 				).permitAll()
-				.anyRequest().authenticated()
-			)
+				.anyRequest().authenticated())
 			.oauth2Login(oauth2 -> oauth2
-				.userInfoEndpoint(userInfo -> userInfo.userService(oAuthService))
+				.userInfoEndpoint(userInfo -> userInfo
+					.userService(oAuthService))
 				.successHandler(oAuth2SuccessHandler)
 				.authorizationEndpoint(authorization -> authorization
 					.authorizationRequestRepository(httpCookieOAuth2AuthorizationRequestRepository)
@@ -106,22 +91,10 @@ public class SecurityConfig {
 	@Bean
 	public CorsConfigurationSource corsConfigurationSource() {
 		CorsConfiguration configuration = new CorsConfiguration();
-		configuration.setAllowedOriginPatterns(List.of(
-			"https://nebulazone-*-uguls-projects.vercel.app",
-			"https://nebulazone-fe.vercel.app",
-			"https://*.nebulazone.store"
-		));
-		configuration.setAllowedMethods(List.of(
-			"GET", "POST", "PUT", "DELETE", "OPTIONS", "PATCH"
-		));
-		configuration.setAllowedHeaders(List.of(
-			"Authorization",
-			"Content-Type",
-			"X-Requested-With"
-		));
-		configuration.setExposedHeaders(List.of(
-			"X-Total-Count"
-		));
+		configuration.setAllowedOriginPatterns(List.of("*"));
+		configuration.setAllowedMethods(List.of("*"));
+		configuration.setAllowedHeaders(List.of("*"));
+		configuration.setExposedHeaders(List.of("*"));
 		configuration.setAllowCredentials(true);
 		configuration.setMaxAge(3600L);
